Two Arrays- Larger from End
The program must accept two integers array of size M and N (where M>=N). Then it must compare the elements of both the arrays at the same position starting from the end and print the larger values. When M is not equal to N, M-N elements in the beginning of the M sized array will not have a pair and hence will be printed as such.

Boundary condition
1<=N<=M<=100

input/output
Input:
6 4
10 20 30 40 50 60
99 30 26 100
output:
100 50 40 99 20 10

Input/output
Input
5 5
10 20 30 40 50
99 30 26 10 77
Output:
77 40 30 30 99