Create Matrix Sum- Bottom Right Quadrant
The program must accept an integer matrix of size RxC as the input. The program must print the sum of integers in the bottom-right quadrant of the matrix as the output.
Boundary Condition(s):
2<=R,c<=50
Input Format:
The first line contains R and C separated by a space.
The next R lines each contains C integer separated by a space.
Output Format:
The first line contains the sum of integers in the bottom-right quadrant of the matrix.
Example Input/Ouput 1:
Input:
4 4
1 2 3 4 
5 6 7 8
9 1 2 3 
4 5 6 7
Output:
18
Explanation:
The integers in the bottom-right quadrant of the matrix are 2 3 6 7. So the sum of  those integers is 18(2+3+6+7)
Hence the Output is 18.