Create Toggle Case- Two String Values
The program must accept two string values s1 and s2 are of equal length as the input. The string s1 contains only alphabets and the string s2 contains u or l(in both lower case and upper case). For each alphabet CH in the string s2, the program must toggle the case of the alphabet at the same position in the string s1 based on the following conditions.
- If CH is u or U, the program must convert the corresponding alphabet in s1 to upper case.
- If CH is l or L, the program must convert the corresponding alphabet in s1 to lower case.
Finally, the pogram must print the modified string s1 as the output.

Boundary Condition(s):
1<=Length fo s1,s2<=1000

Input Format:
The first line contains s1.
The second line contains s2.

Output Format:
The first line contains the modified string s1.

Example Input/Output 1:
Input:
book
uLIU

Output:
BooK

Explanation:
The first alphabet in s2 is u. So the first alphabet in s1 must be in the upper case.
Now the s1 becomes Book.
The second alphabet in s2 is L. So the second alphabet in s1 must be in the lower case.
Now the s1 remains the same Book(as o is already in lower case).
The third alphabet in s2 is l. So the third alphabet in s1 must be in the lower case.
Now the s1 remains the same Book (as o is already in lower case).
The fourth alphabet in s2 is U. So the fourth in s1 must be in the upper case.
Now the s1 becomes BooK.
Hence the output in BooK.

Example Input/Output 2:
Input:
GOOOOGLE
lululUUL

Output:
gOoOoGLe